Refactor FPS reporting for looping
With the introduction of replay looping, several assumptions that were
previously made in the FPS info reporting code are no longer valid.
This includes the assumption that there is only one time window where
the state loading/replay takes place, and the assumption that the
measurement frame range, if specified, occurs entirely within one
replay pass. This commit updates the FPS info code in the aim of
restoring its correctness.