Add option to preserve windows across replay loops
With the current replay logic, desktop windows are created/destroyed
along with each surface/swapchain. Thus, when looping is enabled, it
is common that windows are destroyed at the end of each loop, only to
be almost immediately recreated during the subsequent loop. This
results in an obvious visual disruption in the replay.

This commit adds an option to preserve windows created until the very
end of the replay, causing them to subsist across loop boundaries.

There is no real attempt to match previous uses of the window with
subsequent uses, however. In cases where multiple windows are created
and managed within one application, factors such as creation order,
handle id values, etc. may still cause visual "jumps" during the
replay. For now, though, resolving the issue for the by-far standard
case of single-window captures should be sufficient.