Implement WAR for --preserve-windows with Xlib
This commit puts in a workaround to avoid segfaulting with Xlib when
destroying windows after the replay's vkInstance has been destroyed.
The Xlib context already has one WAR in place, but a new one is
necessary due to the addition of the feature that allows the
preservation of windows across replay loops.